Refactored service to fit with Surtidor definitions

+++ Services/IPistaService.cs
@@ -7,6 +7,6 @@ public interface IPistaService
     void FreeSurtidor(Guid id);
     void BlockSurtidor(Guid id);
     void PrefixSurtidor(Guid id, int max);
-    Task<double> FillSurtidorAsync(Guid id, CancellationToken cancellationToken);
+    Task<int> FillSurtidorAsync(Guid id, CancellationToken cancellationToken);
     IEnumerable<ISurtidor> GetStatus();
 }
